adding backup restore steps
The dump is generated with --clean flag but this may cause errors when the object do not exist. I have included absolute drop of the database to start with a clean sheet. 

Backup is done a bit unfortunately - first tar archiving the home directory and then dumping database to this directory. This should be probably changed so the .tar contains also the dump, right?